moved link preview stuff into dedicated classes
